requirements.txt

viam-sdk==0.41.0
Pillow
opencv-python-headless 
matplotlib
typing-extensions
